Added exceptions to forbid periodic functions with a period of 0ms, as it was experienced that this would lead to memory leaks when allocating memory on the heap in a function with a period of 0ms (e.g., if a string was allocated, the underlying char array might not get freed). Confirmed on Android and MacOS, see TimerBase for details.

+++ include/Utilities/Time.hpp
@@ -188,13 +188,13 @@ public:
 	}
 
 	
-	static Time fromUnixTimestamp(uint32_t seconds)
+	static Time fromUnixTimestamp(uint64_t seconds)
 	{
 		return Time::unixEpoch() + std::chrono::seconds(seconds);
 	}
 
 
-	static Time fromUnixTimestampMilliseconds(uint32_t milliSeconds)
+	static Time fromUnixTimestampMilliseconds(uint64_t milliSeconds)
 	{
 		return Time::unixEpoch() + std::chrono::milliseconds(milliSeconds);
 	}
